chore: Update project files and dependencies for maintenance

+++ database/migrations/2025_04_24_225819_create_events_table.php
@@ -21,8 +21,8 @@ return new class extends Migration
             $table->string('city');
             $table->string('country');
             $table->text('address')->nullable();
-            $table->integer('latitude')->nullable();
-            $table->integer('longitude')->nullable();
+            $table->double('latitude')->nullable();
+            $table->double('longitude')->nullable();
             $table->text('description')->nullable();
             $table->string('ticket_url')->nullable();
             $table->boolean('sold_out')->default(false);
